Remove adapter and move logic into domain function
