+++ apps/cli/src/domain/__tests__/data.ts
@@ -1,15 +1,11 @@
 import { mock } from "vitest-mock-extended";
 
 import { Config } from "../../config.js";
-import {
-  DependencyVersionValidator,
-  PackageJsonReader,
-} from "../package-json.js";
+import { PackageJsonReader } from "../package-json.js";
 import { RepositoryReader } from "../repository.js";
 import { ValidationReporter } from "../validation.js";
 
 export const makeMockDependencies = () => ({
-  dependencyVersionValidator: mock<DependencyVersionValidator>(),
   packageJsonReader: mock<PackageJsonReader>(),
   repositoryReader: mock<RepositoryReader>(),
   validationReporter: mock<ValidationReporter>(),
