+++ apps/cli/src/domain/dependencies.ts
@@ -1,12 +1,8 @@
-import {
-  DependencyVersionValidator,
-  PackageJsonReader,
-} from "./package-json.js";
+import { PackageJsonReader } from "./package-json.js";
 import { RepositoryReader } from "./repository.js";
 import { ValidationReporter } from "./validation.js";
 
 export interface Dependencies {
-  dependencyVersionValidator: DependencyVersionValidator;
   packageJsonReader: PackageJsonReader;
   repositoryReader: RepositoryReader;
   validationReporter: ValidationReporter;
