+++ apps/cli/src/domain/repository.ts
@@ -1,4 +1,6 @@
 import { ok, Result } from "neverthrow";
+import coerce from "semver/functions/coerce.js";
+import semverGte from "semver/functions/gte.js";
 
 import { Config } from "../config.js";
 import { Dependencies } from "./dependencies.js";
@@ -9,6 +11,17 @@ export interface RepositoryReader {
   existsTurboConfig(repoRoot: string): Result<boolean, Error>;
   findRepositoryRoot(cwd: string): Result<string, Error>;
 }
+
+const isVersionValid = (version: string, minVersion: string): boolean => {
+  const minAcceptedSemVer = coerce(minVersion);
+  const dependencySemVer = coerce(version);
+
+  if (!minAcceptedSemVer || !dependencySemVer) {
+    return false;
+  }
+
+  return semverGte(dependencySemVer, minAcceptedSemVer);
+};
 
 export const checkPreCommitConfig =
   (monorepoDir: string) =>
@@ -38,14 +51,10 @@ export const checkPreCommitConfig =
 export const checkTurboConfig =
   (monorepoDir: string) =>
   async (
-    dependencies: Pick<
-      Dependencies,
-      "dependencyVersionValidator" | "packageJsonReader" | "repositoryReader"
-    >,
+    dependencies: Pick<Dependencies, "packageJsonReader" | "repositoryReader">,
     { minVersions }: Config,
   ): Promise<ValidationCheckResult> => {
-    const { dependencyVersionValidator, packageJsonReader, repositoryReader } =
-      dependencies;
+    const { packageJsonReader, repositoryReader } = dependencies;
     const checkName = "Turbo Configuration";
 
     const turboResult = repositoryReader.existsTurboConfig(monorepoDir);
@@ -81,19 +90,18 @@ export const checkTurboConfig =
       });
     }
 
-    return dependencyVersionValidator.isValid(
-      turboDependency,
-      minVersions.turbo,
-    )
-      ? ok({
-          checkName,
-          isValid: true,
-          successMessage:
-            "Turbo configuration is present in the monorepo root and turbo dependency is installed",
-        })
-      : ok({
-          checkName,
-          errorMessage: `Turbo version (${turboDependency.version}) is too low. Minimum required version is ${minVersions.turbo}.`,
-          isValid: false,
-        });
+    if (!isVersionValid(turboDependency.version, minVersions.turbo)) {
+      return ok({
+        checkName,
+        errorMessage: `Turbo version (${turboDependency.version}) is too low. Minimum required version is ${minVersions.turbo}.`,
+        isValid: false,
+      });
+    }
+
+    return ok({
+      checkName,
+      isValid: true,
+      successMessage:
+        "Turbo configuration is present in the monorepo root and turbo dependency is installed",
+    });
   };
