+++ apps/cli/src/index.ts
@@ -4,7 +4,6 @@ import { makeCli } from "./adapters/commander/index.js";
 import { makeValidationReporter } from "./adapters/logtape/validation-reporter.js";
 import { makePackageJsonReader } from "./adapters/node/package-json.js";
 import { makeRepositoryReader } from "./adapters/node/repository.js";
-import { makeDependencyVersionValidator } from "./adapters/semver/dependency-version-validator.js";
 import { getConfig } from "./config.js";
 import { Dependencies } from "./domain/dependencies.js";
 
@@ -21,7 +20,6 @@ await configure({
 });
 
 const deps: Dependencies = {
-  dependencyVersionValidator: makeDependencyVersionValidator(),
   packageJsonReader: makePackageJsonReader(),
   repositoryReader: makeRepositoryReader(),
   validationReporter: makeValidationReporter(),
